+++ ios/Plugin/Utils/UtilsFile.swift
@@ -13,10 +13,15 @@ enum UtilsFileError: Error {
     case deleteFileFailed
     case getAssetsDatabasesPathFailed
     case getDatabasesPathFailed
+    case getDatabasesURLFailed
+    case getApplicationPathFailed
+    case getApplicationURLFailed
     case getFileListFailed
     case copyFromAssetToDatabaseFailed
+    case copyFromNamesFailed
 }
 
+// swiftlint:disable type_body_length
 class UtilsFile {
 
     // MARK: - IsFileExist
@@ -57,24 +62,63 @@ class UtilsFile {
 
     class func getFilePath(fileName: String) throws -> String {
         do {
-            let url = try getDatabasesPath()
+            let url = try getDatabasesURL()
                 return url.appendingPathComponent("\(fileName)").path
-        } catch UtilsFileError.getDatabasesPathFailed {
-            print("Error: getDatabasesPath Failed")
+        } catch UtilsFileError.getDatabasesURLFailed {
+            print("Error: getDatabasesURL Failed")
             throw UtilsFileError.getFilePathFailed
         }
     }
 
-    // MARK: - getDatabasesPath
+    // MARK: - getDatabasesURL
 
-    class func getDatabasesPath() throws -> URL {
+    class func getDatabasesURL() throws -> URL {
         if let path: String = NSSearchPathForDirectoriesInDomains(
             .documentDirectory, .userDomainMask, true
         ).first {
             return NSURL(fileURLWithPath: path) as URL
         } else {
+            print("Error: getDatabasesURL did not find the document folder")
+            throw UtilsFileError.getDatabasesURLFailed
+        }
+    }
+
+    // MARK: - getDatabasesPath
+
+    class func getDatabasesPath() throws -> String {
+        if let path: String = NSSearchPathForDirectoriesInDomains(
+            .documentDirectory, .userDomainMask, true
+        ).first {
+            return path
+        } else {
             print("Error: getDatabasesPath did not find the document folder")
             throw UtilsFileError.getDatabasesPathFailed
+        }
+    }
+
+    // MARK: - getApplicationURL
+
+    class func getApplicationURL() throws -> URL {
+        if let path: String = NSSearchPathForDirectoriesInDomains(
+            .applicationDirectory, .userDomainMask, true
+        ).first {
+            return NSURL(fileURLWithPath: path) as URL
+        } else {
+            print("Error: getApplicationURL did not find the application folder")
+            throw UtilsFileError.getApplicationURLFailed
+        }
+    }
+
+    // MARK: - getApplicationPath
+
+    class func getApplicationPath() throws -> String {
+        if let path: String = NSSearchPathForDirectoriesInDomains(
+            .applicationDirectory, .userDomainMask, true
+        ).first {
+            return path
+        } else {
+            print("Error: getApplicationPath did not find the application folder")
+            throw UtilsFileError.getApplicationPathFailed
         }
     }
 
@@ -126,6 +170,29 @@ class UtilsFile {
         }
     }
 
+    // MARK: - CopyFromNames
+
+    class func copyFromNames(dbPathURL: URL, fromFile: String, databaseURL: URL,
+                             toFile: String) throws {
+        do {
+            let uFrom: URL = dbPathURL.appendingPathComponent(fromFile)
+            let uTo: URL = databaseURL.appendingPathComponent(toFile)
+            let pFrom: String = uFrom.path
+            let pTo: String = uTo.path
+            let bRet: Bool = try copyFile(pathName: pFrom, toPathName: pTo)
+            if bRet {
+                return
+            } else {
+                print("Error: FromNames return false")
+                throw UtilsFileError.copyFromNamesFailed
+            }
+        } catch UtilsFileError.copyFileFailed {
+            print("Error: copyFile Failed")
+            throw UtilsFileError.copyFromNamesFailed
+        }
+
+    }
+
     // MARK: - CopyFromAssetToDatabase
 
     class func copyFromAssetToDatabase(fromDb: String, toDb: String) throws {
@@ -133,7 +200,7 @@ class UtilsFile {
             let uAsset: URL = try getAssetsDatabasesPath()
                                         .appendingPathComponent(fromDb)
             let pAsset: String = uAsset.path
-            let uDb: URL = try getDatabasesPath()
+            let uDb: URL = try getDatabasesURL()
                                         .appendingPathComponent(toDb)
             let pDb: String = uDb.path
             let bRet: Bool = try copyFile(pathName: pAsset, toPathName: pDb)
@@ -146,8 +213,8 @@ class UtilsFile {
         } catch UtilsFileError.getAssetsDatabasesPathFailed {
             print("Error: getAssetsDatabasesPath Failed")
             throw UtilsFileError.copyFromAssetToDatabaseFailed
-        } catch UtilsFileError.getDatabasesPathFailed {
-            print("Error: getDatabasesPath Failed")
+        } catch UtilsFileError.getDatabasesURLFailed {
+            print("Error: getDatabasesURL Failed")
             throw UtilsFileError.copyFromAssetToDatabaseFailed
         } catch UtilsFileError.copyFileFailed {
             print("Error: copyFile Failed")
@@ -242,6 +309,23 @@ class UtilsFile {
         return ret
     }
 
+    // MARK: - DeleteFile
+
+    class func deleteFile(dbPathURL: URL, fileName: String) throws -> Bool {
+        var ret: Bool = false
+        do {
+            let uURL: URL = dbPathURL.appendingPathComponent(fileName)
+            let filePath: String = uURL.path
+            ret = try deleteFile(filePath: filePath)
+        } catch UtilsFileError.deleteFileFailed {
+            throw UtilsFileError.deleteFileFailed
+        } catch let error {
+            print("Error: \(error)")
+            throw UtilsFileError.deleteFileFailed
+        }
+        return ret
+    }
+
     // MARK: - RenameFile
 
     class func renameFile (filePath: String, toFilePath: String)
@@ -261,3 +345,4 @@ class UtilsFile {
         }
     }
 }
+// swiftlint:enable type_body_length